Add check for existing icons before downloading and setting them
This commit adds a check to see if the icons already exist before downloading them. If the icons are not present, they will be downloaded and set as the recycle bin icons. If there is an error during the
download or setting of the icons, an error message will be displayed.